refactor: 로그인 파일 구조 및 네이밍 변경
