--- src/components/login-modal/index.ts
@@ -1,1 +1,0 @@
-export { default as LoginModal } from './LoginModal';
