+++ src/pages/login/index.ts
@@ -1,2 +1,1 @@
-export { default as LoginButton } from './LoginButton';
 export { default as KakaoLoginPage } from './Login';
